Restructured directory
Restructured directory layout and organized files
