+++ src/models/Cliente.ts
@@ -1,4 +1,6 @@
-import { GeneradorID } from "./GeneradorID";
+//TODO: FACTORY METHOD para crear clientes. veterinariaFactory.crearCliente("Juan Perez", 12345678);
+
+import { GeneradorID } from "../app/GeneradorID";
 
 export class Cliente {
   private id: number;
