+++ src/repositories/Repository.ts
@@ -1,0 +1,1 @@
+//abstract con los metodos CRUD
